Move Select from classical to photonic
